+++ R/remove_env.R
@@ -7,6 +7,15 @@
 #' @export
 remove_env <- function(env_name = "condathis-env",
                        verbose = "silent") {
+  if (isFALSE(env_exists(env_name))) {
+    cli::cli_abort(
+      message = c(
+        `x` = "Environment {.field {env_name}} do not exist.",
+        `!` = "Check {.code list_envs()} for available environments."
+      ),
+      class = "condathis_error_env_remove"
+    )
+  }
   px_res <- native_cmd(
     conda_cmd = "env",
     conda_args = c(
@@ -18,5 +27,12 @@ remove_env <- function(env_name = "condathis-env",
     ),
     verbose = verbose
   )
+  if (isTRUE(verbose %in% c("full", "output"))) {
+    cli::cli_inform(
+      message = c(
+        `!` = "Environment {.field {env_name}} succesfully removed."
+      )
+    )
+  }
   return(invisible(px_res))
 }