+++ R/list_packages.R
@@ -26,6 +26,16 @@
 #' }
 #' @export
 list_packages <- function(env_name = "condathis-env", verbose = "silent") {
+  if (isFALSE(env_exists(env_name))) {
+    cli::cli_abort(
+      message = c(
+        `x` = "Environment {.field {env_name}} do not exist.",
+        `!` = "Check {.code list_envs()} for available environments."
+      ),
+      class = "condathis_list_packages_missing_env"
+    )
+  }
+
   px_res <- native_cmd(
     conda_cmd = "list",
     conda_args = c(